requirements.txt
# Django
Django==5.2.3
django-admin-interface==0.30.1
django-tinymce==4.1.0
django-advance-thumbnail==1.0.2
django-health-check==3.19.2
django-cleanup==9.0.0
django-select2==8.4.1
django-ninja==1.4.3
django-ninja-extra==0.30.1
django-modelclone-next==0.8.2
django-vite==3.1.0
django-bootstrap5==25.1
django-vectortiles==1.0.1
django-leaflet==0.32.0
django-cache-cleaner==0.2.0

# Environment variables
django-environ==0.12.0
python-dotenv==1.1.0

# prod server
gunicorn==23.0.0

# image manipulation
Pillow==11.2.1

# faster django-ninja json (for geojson)
orjson==3.10.18

# Dataabse (psycopg[binary] cannot be locked to specific version number)
psycopg[binary]
